+++ tests/TestCase.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace NotificationChannels\ItsChats\Tests;
+namespace NotificationChannels\Gronosync\Tests;
 
-use NotificationChannels\ItsChats\ItsChatsServiceProvider;
+use NotificationChannels\Gronosync\GronosyncServiceProvider;
 use Orchestra\Testbench\TestCase as Orchestra;
 
 abstract class TestCase extends Orchestra
@@ -12,14 +12,14 @@ abstract class TestCase extends Orchestra
     protected function getPackageProviders($app): array
     {
         return [
-            ItsChatsServiceProvider::class,
+            GronosyncServiceProvider::class,
         ];
     }
 
     protected function getEnvironmentSetUp($app): void
     {
-        $app['config']->set('services.itschats', [
-            'url'   => 'https://itschats.test',
+        $app['config']->set('services.gronosync', [
+            'url'   => 'https://api.gronosync.test',
             'token' => 'test-token-123',
         ]);
     }